```
from asv_bench.benchmarks.io.stata import *

self = StataMissing()
self.setup("tc")

%timeit self.time_read_stata("tc")
150 ms ± 3.3 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
114 ms ± 2.81 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR
```

cc @bashtage